# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hush Service API
  version: 1.0.0
  description: Single-owner encrypted vault. CLI devices are read-only. All binary
    fields use canonical unpadded base64url. See crypto.md for mandatory AAD,
    algorithm, size, and nonce rules. The Worker validates structure, not
    cryptographic correctness.
servers:
  - url: /api/v1
paths:
...
  /database:
    get:
      operationId: listDatabaseTables
      summary: List inspectable application tables (owner only)
      security: *a3
      parameters: []
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                type: object
                properties:
                  tables:
                    type: array
                    items:
                      type: string
                      enum: &a5
                        - vault
                        - secrets
                        - profiles
                        - devices
                        - device_activity
                        - audit_log
                        - secret_nonces
                        - master_nonces
                required:
                  - tables
                additionalProperties: false
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /database/{table}:
    get:
      operationId: getDatabaseRows
      summary: Read raw SQLite rows; never decrypts or executes supplied SQL
      security: *a3
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
            enum: *a5
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 1000000
            default: 0
      responses:
        "200":
          description: Success. Never contains plaintext secrets or private keys.
          headers:
            ETag:
              description: Quoted vault revision when a vault exists.
              schema: *a2
            Cache-Control:
              schema:
                const: no-store
          content:
            application/json:
              schema:
                type: object
                properties:
                  table: *a2
                  columns:
                    type: array
                    items:
                      type: object
                      properties:
                        cid:
                          type: integer
                        name: *a2
                        type: *a2
                        notnull:
                          type: integer
                        dflt_value: {}
                        pk:
                          type: integer
                        hidden:
                          type: integer
                      required:
                        - cid
                        - name
                        - type
                        - notnull
                        - dflt_value
                        - pk
                        - hidden
                      additionalProperties: false
                  rows:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  total:
                    type: integer
                    minimum: 0
                  offset:
                    type: integer
                    minimum: 0
                  limit:
                    const: 50
                required:
                  - table
                  - columns
                  - rows
                  - total
                  - offset
                  - limit
                additionalProperties: false
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid Access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Unauthorized identity, revoked device, or invalid admin origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Revision, uniqueness, nonce reuse, or capacity conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: JSON required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "428":
          description: If-Match required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Access not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *a2
            request_id: *a2
          required:
            - code
            - request_id
          additionalProperties: false
      required:
        - error
      additionalProperties: false
  securitySchemes:
    AccessSession:
      type: apiKey
      in: cookie
      name: CF_Authorization
      description: Cloudflare Access human session. Edge forwards the signed
        assertion; Worker verifies it and matches OWNER_EMAIL.
    AccessClientId:
      type: apiKey
      in: header
      name: CF-Access-Client-Id
    AccessClientSecret:
      type: apiKey
      in: header
      name: CF-Access-Client-Secret